+++ source/archives.c
@@ -44,7 +44,7 @@ extern "C" {
 #include <malloc.h>
 #include <sys/stat.h>
 
-#include "Archives.h"
+#include "archives.h"
 
 /******************************************************************************
  ** UnZip *********************************************************************
