+++ source/cia.h
@@ -1,0 +1,6 @@
+// KernelTimeMachine
+// Safe CIA manager
+// Licensed under GNU General Public License (GPL)
+// Check out https://www.gnu.org/licenses/gpl.html
+
+bool installCIA(char *path,u8 mediatype,u64 *installedTitleIDs,char *name);
