+++ source/menu.h
@@ -1,0 +1,10 @@
+// KernelTimeMachine
+// Safe CIA manager
+// Licensed under GNU General Public License (GPL)
+// Check out https://www.gnu.org/licenses/gpl.html
+
+u8 mainMenu();
+u8 downgradeMenu();
+
+extern PrintConsole botConsole;
+extern PrintConsole topConsole;
