+++ source/utils.h
@@ -1,0 +1,11 @@
+// KernelTimeMachine
+// Safe CIA manager
+// Licensed under GNU General Public License (GPL)
+// Check out https://www.gnu.org/licenses/gpl.html
+
+int error(char *msg,u8 errorCode);
+void clearScreen();
+
+extern u8 *fbBottom;
+extern u8 *fbTopRight;
+extern u8 *fbTopLeft;
